extract connection list to App
