--- src/screens/database/components/ConnectionList/style.module.scss
@@ -1,19 +1,0 @@
-.connection-options {
-	display: none;
-}
-
-.connection:hover .connection-options {
-	display: flex;
-}
-
-.group {
-	.group-remove {
-		display: none;
-		position: absolute;
-		right: var(--mantine-spacing-xl);
-	}
-
-	&:hover .group-remove {
-		display: block;
-	}
-}
